Make build-script more Cargo-friendly
Previously, the build script would write generated code into the `src`
directory, which is not officially supported: when one depends on
mime_guess, the src dir is somewhere in `.cargo/registry`, and is
supposed to be immutable.

The correct way is to gen stuff into `OUT_DIR`, and this is exactly what
this commit is doing.

Additionally, it moves the build file to the now-default `build.rs`.

closes #31

(cherry picked from commit 1e1ce89ce37373617ac9ff8a6fd6a680bbe7a454)

# Conflicts:
#	src/lib.rs